Program was updated.

+++ src/main/java/pl/patrycja/game/Main.java
@@ -1,12 +1,11 @@
 package pl.patrycja.game;
 
-import pl.patrycja.game.doubleValues.DoubleRange;
 import pl.patrycja.game.integerValues.IntegerRange;
 
 public class Main {
 
     public static void main(String[] args) {
-        Range range = new DoubleRange(1.0, 10.0);
+        Range range = new IntegerRange(1, 10);
         System.out.println("Hello in my magic game. Guess the number between " + range + ". Enjoy!");
         Game game = new Game(range);
         game.play();
